Nightly builds You can find builds for pretty much any recent Cardinal commit here. Just click on any successful build, and scroll to the bottom to find the builds. (note the canvas-like area in the middle prevents mouse wheel scrolling)
A GitHub account is required in order to download these builds.

Building
Basic building instructions are available in BUILDING.md.

Plugin variants
Cardinal provides 3 plugin variants - "main", Synth and FX.
They are all equivalent in performance and behaviour, with only the IO and metadata that changes.
This is because some hosts are very strict on which plugins are allowed as instruments vs FX, so separate variants of the same plugin are needed.

FX and Synth variants both have 2 audio outputs, while "main" has 8.
All variants have MIDI input and output support.

Main
This variant provides 8 audio inputs and outputs and 10 CV inputs and outputs.

NOTE: Due to AU and VST2 formats not supporting CV ports, this variant is not available for those formats.

Synth
This variant provides 2 audio outputs but no audio inputs or CV ports.
Plugin type is set as "instrument".

FX
This variant provides 2 audio inputs and outputs, but no CV ports.
Plugin type is set as regular "ef